Move orion sink parsing to misc
