+++ yt/frontends/boxlib/misc.py
@@ -2,8 +2,72 @@ import os
 
 import numpy as np
 
-from yt.frontends.chombo.io import parse_orion_sinks
+from yt.funcs import mylog
 from yt.geometry.selection_routines import GridSelector
+
+
+def parse_orion_sinks(fn):
+    r"""
+    Orion sink particles are stored in text files. This function
+    is for figuring what particle fields are present based on the
+    number of entries per line in the \*.sink file.
+    """
+
+    # Figure out the format of the particle file
+    with open(fn) as f:
+        lines = f.readlines()
+
+    try:
+        line = lines[1]
+    except IndexError:
+        # a particle file exists, but there is only one line,
+        # so no sinks have been created yet.
+        index = {}
+        return index
+
+    # The basic fields that all sink particles have
+    index = {
+        "particle_mass": 0,
+        "particle_position_x": 1,
+        "particle_position_y": 2,
+        "particle_position_z": 3,
+        "particle_momentum_x": 4,
+        "particle_momentum_y": 5,
+        "particle_momentum_z": 6,
+        "particle_angmomen_x": 7,
+        "particle_angmomen_y": 8,
+        "particle_angmomen_z": 9,
+        "particle_id": -1,
+    }
+
+    if len(line.strip().split()) == 11:
+        # these are vanilla sinks, do nothing
+        pass
+
+    elif len(line.strip().split()) == 17:
+        # these are old-style stars, add stellar model parameters
+        index["particle_mlast"] = 10
+        index["particle_r"] = 11
+        index["particle_mdeut"] = 12
+        index["particle_n"] = 13
+        index["particle_mdot"] = 14
+        index["particle_burnstate"] = 15
+
+    elif len(line.strip().split()) == 18 or len(line.strip().split()) == 19:
+        # these are the newer style, add luminosity as well
+        index["particle_mlast"] = 10
+        index["particle_r"] = 11
+        index["particle_mdeut"] = 12
+        index["particle_n"] = 13
+        index["particle_mdot"] = 14
+        index["particle_burnstate"] = 15
+        index["particle_luminosity"] = 16
+    else:
+        # give a warning if none of the above apply:
+        mylog.warning("Warning - could not figure out particle output file")
+        mylog.warning("These results could be nonsense!")
+
+    return index
 
 
 class BoxlibParticleSelectionMixin:
